client - adding components

+++ Domain/Entitites/Product.cs
@@ -12,7 +12,7 @@ namespace Domain.Entitites
 
         public int Stock { get; set; }
 
-        public long Price { get; set; }
+        public float Price { get; set; }
 
         public ICollection<Order> Orders { get; set; }
     }
